move url check-urlCheck.js, add duplication check with houses in local storage
